requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
drf-spectacular
psycopg2-binary
redis
django-redis
celery
django-celery-beat
django-celery-results
web3
stripe
google-generativeai
pydantic
python-decouple
python-dotenv
Pillow
requests
scipy
scikit-learn
numpy
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-text-splitters
google-generativeai
faiss-cpu
sentence-transformers